pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "deckdown"
version = "0.1.0"
description = "A CLI-based presentation framework that renders Markdown slides in the terminal"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
dependencies = [
    "rich>=13.0",
    "textual>=0.40",
    "python-frontmatter>=1.0",
    "click>=8.0",
    "markdown-it-py>=3.0",
]

[project.scripts]
deckdown = "deckdown.cli:main"

[tool.setuptools.packages.find]
where = ["src"]
